# Models/Configuracion.cs
namespace Periodico.Models
{
    public class Configuracion
    {
        public int Id { get; set; }
        public string Nombre { get; set; }
        public string Correo { get; set; }
        public int Telefono { get; set; }
        public string Direccion { get; set; }
        public string Estado { get; set; }
    }
}