# bin/Release/net8.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Periodico, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Categorias/{id}": {
      "get": {
        "tags": [
          "Categorias"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Categoria"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Categorias"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Categorias/Actualizar": {
      "put": {
        "tags": [
          "Categorias"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "descripcion",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Categorias/Crear": {
      "post": {
        "tags": [
          "Categorias"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "descripcion",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/Configuraciones/{id}": {
      "get": {
        "tags": [
          "Configuraciones"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Configuracion"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Configuraciones"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Configuraciones/Actualizar": {
      "put": {
        "tags": [
          "Configuraciones"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "nombre",
            "type": "string"
          },
          {
            "in": "query",
            "name": "correo",
            "type": "string"
          },
          {
            "in": "query",
            "name": "telefono",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "direccion",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Configuraciones/Crear": {
      "post": {
        "tags": [
          "Configuraciones"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "nombre",
            "type": "string"
          },
          {
            "in": "query",
            "name": "correo",
            "type": "string"
          },
          {
            "in": "query",
            "name": "telefono",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "direccion",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Categoria": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "descripcion": {
          "type": "string"
        },
        "estado": {
          "type": "string"
        },
        "publicaciones": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Publicacion"
          }
        }
      }
    },
    "Configuracion": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "nombre": {
          "type": "string"
        },
        "correo": {
          "type": "string"
        },
        "telefono": {
          "format": "int32",
          "type": "integer"
        },
        "direccion": {
          "type": "string"
        },
        "estado": {
          "type": "string"
        }
      }
    },
...
    "Publicacion": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "titulo": {
          "type": "string"
        },
        "fecha": {
          "format": "date-time",
          "type": "string"
        },
        "descripcion": {
          "type": "string"
        },
        "imagen": {
          "type": "string"
        },
        "visualizacion": {
          "format": "int32",
          "type": "integer"
        },
        "estado": {
          "type": "string"
        },
        "idusuario": {
          "format": "int32",
          "type": "integer"
        },
        "idcategoria": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
...
  }
}